+++ src/interpreter/InterpreterState.ts
@@ -5,12 +5,14 @@
 
 import {SimpleCellAddress} from '../Cell'
 import {FormulaVertex} from '../DependencyGraph/FormulaVertex'
+import {ActiveEdgeCollector} from './ActiveEdgeCollector'
 
 export class InterpreterState {
   constructor(
     public formulaAddress: SimpleCellAddress,
     public arraysFlag: boolean,
     public formulaVertex?: FormulaVertex,
+    public activeEdgeCollector?: ActiveEdgeCollector,
   ) {
   }
 }